requirements.txt
aiobotocore==2.7.0
aiohttp==3.9.0
aioitertools==0.11.0
aiosignal==1.3.1
anndata==0.9.2
anyio==4.1.0
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.28.64
botocore==1.31.64
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
croniter==1.3.15
cycler==0.12.1
datasets==2.14.4
dateutils==0.6.12
deepdiff==6.7.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.88.0
filelock==3.13.1
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.1
inquirer==3.1.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lightning==1.9.0
lightning-cloud==0.5.55
lightning-utilities==0.10.0
llvmlite==0.39.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.56.4
numpy==1.23.5
ordered-set==4.1.0
packaging==22.0
pandas==2.0.3
pathtools==0.1.2
patsy==0.5.3
Pillow==9.3.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pydantic==1.10.13
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0
readchar==4.0.5
requests==2.31.0
rich==13.7.0
s3fs==2023.10.0
s3transfer==0.7.0
safetensors==0.4.0
scanpy==1.9.4
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.36.0
session-info==1.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
somacore==1.0.4
soupsieve==2.5
starlette==0.22.0
starsessions==1.3.0
statsmodels==0.14.0
stdlib-list==0.10.0
threadpoolctl==3.2.0
tiledb==0.22.3
timm==0.9.6
torchmetrics==1.2.0
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.5
urllib3==1.26.18
uvicorn==0.24.0.post1
wandb==0.15.12
wcwidth==0.2.12
websocket-client==1.6.4
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
